Third version
This is my third version of the code. I refactored part A to add the functionality of generating an 8-digit unique booking code, which is stored in the seat_map instead of using “R”. At the same time, the booking code is recorded in the passenger_records dictionary to facilitate name-based queries.